Move HttpFilter to the correct package. Re-order some methods to simplify checking against the RI using javap.
git-svn-id: https://svn.apache.org/repos/asf/tomcat/trunk@1707821 13f79535-47bb-0310-9956-ffa450edef68
